Drop bytecode generation from the MethodHandle implementation.
The ideal way to invoke a method handle is by binding it to an `invokedynamic` instruction.  This provides the VM with a lot of information about the lifetime and target of the methodhandle.  On the other hand calling `invokeExact` on a more 'dynamic' methodhandle is slower since the VM always needs to do things like verify that the signature of the callsite matches the handle, and then resolve to the actual target.  See https://wiki.openjdk.org/display/HotSpot/Deconstructing+MethodHandles for details.

This is why the methodhandle implementation initially pursued bytecode generation, simply so we could bind to an `invokedynamic` instruction.  However, this comes with a massive downside of needing to generate tons of classes.  This slows down injector creation substantially and puts pressure on the vms `Compressed class space` in larger applications which limits scalability of this solution.  So mitigate those risks this commit undoes the bytecode generation.

A compromise solution would probably be some kind of `JIT` approch where we generate a custom provider once it gets hot enough so we would only bother with those Provider instances that get invoked a lot.  However, im guessing the small perf penalty of `invokeExact` on a non-constant handle will be small and fixed and such it probably won't be too impactful in the limit.