Add support for caching MethodHandles.
Each method handle is potentially compiled to a concrete java method and
so we want to avoid recreating identical handles.  This should help
preserve VM resources like 'compressed class space' and improve
linkage performance.  The caching strategy is a little complicated since
we need to respect the `linked` boolean parameter.  Most factories
ignore this parameter but some respect it so our caching strategy allows
for both.