Add MethodHandle support for static member injection.
How valuable this is is somewhat debatable, but it is necessary if we want to make the fast-class support completely optional in the future.

This revealed a number of latent issues I am addressing at the same time.  Notably, some of the `asCollector` calls used for the fast-class fallbacks were not correct so this just eliminates all of them.