Simplify the logic for construct large `Object[]`, `ImmutableMap`, and `ImmutableSet` objects
Some investigation in https://stackoverflow.com/questions/79551257/efficiently-build-a-large-array-with-methodhandles revealed that the simple recursion approach was often the best so all approaches were simplified to do that.  For the ImmutableMap/ImmutableSet cases, special cases were added for the `of` overloads that are supported by guava. From perusing the source these trigger fast paths and allow us to avoid the `Builder` allocation.